+++ TG.js
@@ -16,6 +16,7 @@ class TG {
 
   /*
    *  Sends text message to a existing chat.
+   *  Method "parseTextEntities" requires TDLib 1.1.0 (git.io/tdlibchanges).
    */
   async sendTextMessage(chatId, text, options = {}) {
     let formattedText
@@ -58,6 +59,55 @@ class TG {
     }
     combine(payload, options)
     return this.client.fetch(payload)
+  }
+
+  /*
+   *  Get all chats.
+   */
+  async getAllChats() {
+    let { chat_ids: chats} = await this.client.fetch({
+      '@type': 'getChats', 
+      'offset_order': '9223372036854775807',
+      'offset_chat_id': 0,
+      'limit': 99999,
+    })
+    chats = chats.map((chatId) => this.client.fetch({
+      '@type': 'getChat', 
+      'chat_id': chatId,
+    }))
+    chats = await Promise.all(chats)
+    return chats
+  }
+
+  /*
+   *  Update user's username.
+   *  Method "checkChatUsername" requires TDLib 1.2.0 (git.io/tdlibchanges).
+   */
+  updateUsername(username, supergroupId = null) {
+    return new Promise(async (resolve, reject) => {
+      let payload
+      // Check the username
+      payload = { '@type': 'checkChatUsername', username }
+      if (supergroupId) {
+        const { id } = await this.client.fetch({ '@type': 'createSupergroupChat', 'supergroup_id': supergroupId })
+        payload['chat_id'] = id
+      } else {
+        const { id } = await this.client.fetch({ '@type': 'getMe' })
+        payload['chat_id'] = id
+      }
+      const { '@type': checkResult } = await this.client.fetch(payload)
+      if (checkResult !== 'checkChatUsernameResultOk') {
+        return reject(checkResult)
+      }
+      // Update the username
+      payload = { '@type': 'setUsername', username }
+      if (supergroupId) {
+        payload['@type'] = 'setSupergroupUsername'
+        payload['supergroup_id'] = supergroupId
+      }
+      const result = await this.client.fetch(payload)
+      resolve(result)
+    })
   }
 }